requirements.txt
asttokens==3.0.1
async-timeout==5.0.1
attrs==26.1.0
backports.zstd==1.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
brotli==1.2.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
debugpy==1.8.0
decorator==5.2.1
executing==2.2.1
flasgger==0.9.7.1
Flask==3.1.1
Flask-Compress==1.23
Flask-Cors==5.0.0
Flask-DebugToolbar==0.16.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mistune==3.2.0
packaging==25.0
parso==0.8.6
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.12.1
pytest==8.4.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==5.3.1
referencing==0.37.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.6.0
Werkzeug==3.1.3
